+++ src/validations/auth.js
@@ -1,3 +1,21 @@
+const validateLoginForm = (values) => {
+  const errors = {};
+
+  const trimmedLogin = values.login?.trim();
+  if (!trimmedLogin) {
+    errors.login = "Obrigatório";
+  } else if (trimmedLogin.length < 3) {
+    errors.login = "O login digitado é muito curto";
+  }
+
+  if (!values.password) {
+    errors.password = "Obrigatório";
+  } else if (values.password.length < 4) {
+    errors.password = "A senha digitada é muito curta";
+  }
+
+  return errors;
+};
 const passwordValidation = (values) => {
   const errors = {};
 
@@ -33,4 +51,4 @@ const passwordValidation = (values) => {
   return errors;
 };
 
-export { passwordValidation };
+export { validateLoginForm, passwordValidation };
