+++ src/validations/index.js
@@ -1,8 +1,4 @@
-export { default as validateLoginForm } from "./LoginValidations";
-export { validateUserCreateForm } from "./UserValidations";
-export { validateUserEditForm } from "./UserValidations";
-export { validateContacts } from "./UserValidations";
-export { validateCPForCNPJ } from "./UserValidations";
-export { validateDepartmenCreateForm } from "./DepartmentValidations";
-export { validateDepartmenEditForm } from "./DepartmentValidations";
-export { validateConstructionCreateForm } from "./ConstructionValidations";
+export * from "./auth";
+export * from "./construction";
+export * from "./department";
+export * from "./user";
